Split AssignJobDialog mutation logic into dedicated hook
